backend/requirements.txt
Flask==2.2.2
flask-cors==4.0.1
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==9.4.0